More space-efficient export of uvs
Possibly faster on average since numpy.unique will use 'quicksort' sorting.

Might be slower due to the arrangement of uvs within uv layers being faster with the sorting algorithm chosen for stable sorting when return_index is True (the majority of the time is spent on the sorting done by numpy.unique).